11/26

+++ app/src/main/java/com/itcast/iraqishome/bean/LimitedDetailBean.java
@@ -29,7 +29,7 @@ public class LimitedDetailBean {
     public class ItemInfo{
         public float ActivityPrice;
         public String Appeal;
-        public int Code;
+        public long Code;
         public int CommentCount;
         public String ImageUrl;
         public int ItemInfoId;
